+++ src/directives/gridContainerTemplate/gridContainerTemplate.directive.ts
@@ -1,5 +1,7 @@
 import {Directive, TemplateRef, inject} from '@angular/core';
 
+import {GridContext} from '../../interfaces';
+
 /**
  * Directive used for obtaining template for grid content renderer container
  */
@@ -10,8 +12,21 @@ import {Directive, TemplateRef, inject} from '@angular/core';
 })
 export class GridContainerTemplateSADirective
 {
+    //######################### public properties #########################
+
     /**
      * Obtained template by this directive
      */
-    public template: TemplateRef<unknown> = inject(TemplateRef);
+    public template: TemplateRef<GridContext> = inject(TemplateRef<GridContext>);
+
+    //######################### ng language server #########################
+    
+    /**
+     * Allows typechecking for template
+     */
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    static ngTemplateContextGuard(_dir: GridContainerTemplateSADirective, _ctx: unknown): _ctx is GridContext<any>
+    {
+        return true;
+    }
 }