+++ src/interfaces/gridColumn/index.ts
@@ -1,2 +1,3 @@
 export * from './gridColumn.interface';
+export * from './matrixGridColumn.interface';
 export * from './tableGridColumn.interface';